Fixed bug in Item interface when comparing (0,0) for floating point iimplementations, completed Exercise 4.25

+++ Chapter4/PushdownStack/Examples/Program4_5/stackList.c
@@ -48,7 +48,7 @@ STACKnode* NEW(Item item, STACKnode* next) {
     return x;
 }
 
-void STACKinit(int maxN) {
+void STACKinit(size_t maxN) {
     //reset stack
     head = nullptr;
 }
